refactor: rename provider file to match class names
